i18n: Move the babel plugin to a separate package
