+++ packages/babel-plugin-i18n/.npmrc
@@ -1,0 +1,1 @@
+package-lock=false
