+++ packages/babel-plugin-i18n/README.md
@@ -1,0 +1,18 @@
+i18n
+======
+
+Babel plugin used to scan JavaScript files for use of localization functions. It then compiles these into a [gettext POT formatted](https://en.wikipedia.org/wiki/Gettext) file as a template for translation. By default the output file will be written to `gettext.pot` of the root project directory. This can be overridden using the `"output"` option of the plugin.
+
+```json
+[ "@wordpress/babel-plugin-i18n", {
+	"output": "languages/myplugin.pot"
+} ]
+```
+
+## Installation
+
+Install the module:
+
+```bash
+npm install @wordpress/babel-plugin-i18n --save-dev
+```
