+++ packages/babel-plugin-i18n/package.json
@@ -1,13 +1,15 @@
 {
-	"name": "@wordpress/i18n",
+	"name": "@wordpress/babel-plugin-i18n",
 	"version": "1.0.0-beta.0",
-	"description": "WordPress i18n library",
+	"description": "WordPress Babel i18n Plugin",
 	"author": "WordPress",
 	"license": "GPL-2.0-or-later",
 	"keywords": [
-		"i18n"
+		"i18n",
+		"babel",
+		"plugin"
 	],
-	"homepage": "https://github.com/WordPress/packages/tree/master/packages/i18n/README.md",
+	"homepage": "https://github.com/WordPress/packages/tree/master/packages/babel-plugin-i18n/README.md",
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/WordPress/packages.git"
@@ -15,15 +17,16 @@
 	"bugs": {
 		"url": "https://github.com/WordPress/packages/issues"
 	},
-	"main": "build/index.js",
-	"module": "build-module/index.js",
+	"main": "src/index.js",
 	"publishConfig": {
 		"access": "public"
 	},
 	"dependencies": {
 		"gettext-parser": "^1.3.1",
-		"jed": "^1.1.1",
-		"lodash": "^4.17.5",
-		"memize": "^1.0.5"
+		"lodash": "^4.17.5"
+	},
+	"devDependencies": {
+		"babel-core": "^6.26.0",
+		"babel-traverse": "^6.26.0"
 	}
 }
