+++ packages/babel-plugin-i18n/src/index.js
